Group4_BloodCancerDetection/requirements.txt
# Blood Cancer Detection System - Requirements
# Python 3.13 Compatible

# Deep Learning & Computer Vision
tensorflow>=2.15.0
keras>=3.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Machine Learning Models
scikit-learn>=1.3.0
xgboost>=2.0.0

# Grad-CAM Visualization
tf-keras-vis>=0.8.0

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
